bug fix in telemetrydefinition

+++ onboard/source/core/include/TelemetryDefinition.hh
@@ -65,7 +65,7 @@ public:
   void setChamberPressure(uint16_t v) { chamberPressure_ = v; }
   void setChamberPressureNEU(int index, float v){ if (index<static_cast<int>(chamberPressureNEU_.size())) chamberPressureNEU_[index] = v; }
   void resizeChamberPressureNEU(int n) { chamberPressureNEU_.resize(n); }
-  void setJacketPressureNEU(int index, int32_t v){ if (index<static_cast<int>(jacketPressureNEU_.size())) jacketPressureNEU_[index] = v; }
+  void setJacketPressureNEU(int index, float v){ if (index<static_cast<int>(jacketPressureNEU_.size())) jacketPressureNEU_[index] = v; }
   void resizeJacketPressureNEU(int n) { jacketPressureNEU_.resize(n); }
   void setCompressorTemperature(int index, int16_t v){ if (index<static_cast<int>(compressorTemperature_.size())) compressorTemperature_[index] = v; }
   void resizeCompressorTemperature(int n) { compressorTemperature_.resize(n); }
